Make SimpleList configurable
